Rename friktion package name
friktion_swap_client -> friktion
Keep the standard of other clients.
