+++ friktion/friktion/constants.py
