+++ friktion/friktion/inertia_anchor/accounts/stub_oracle.py
