+++ friktion/friktion/inertia_anchor/instructions/create_stub_oracle.py
