+++ friktion/friktion/inertia_anchor/instructions/set_stub_oracle.py
