+++ friktion/friktion/pda.py
@@ -1,9 +1,10 @@
 from typing import Tuple
 
-from friktion_swap_client.friktion_anchor.accounts.user_orders import UserOrders
-from friktion_swap_client.friktion_anchor.program_id import PROGRAM_ID
 from solana.publickey import PublicKey
 from solana.rpc.async_api import AsyncClient
+
+from friktion.friktion_anchor.accounts.user_orders import UserOrders
+from friktion.friktion_anchor.program_id import PROGRAM_ID
 
 
 def find_swap_order_address(user: PublicKey, order_id: int) -> Tuple[PublicKey, int]:
